TelegramBot/requirements.txt
aiogram~=3.15.0
python-dotenv~=1.0.1
requests~=2.32.3
g4f~=0.3.8.3